Update all tests for contract changes
